Resolve preloadings in mappings
Currently, it is not possible to set a parameter to resolve to a mapping:

```
dp.from = df.now
```

It does not get translated to a real time. It remains as "df.now"

This translation is interesting for data recovery. We can parameterize
prunning of bronze tables with default arguments, like

```
dp.from = df.start_of_today
dp.to = df.now
```

When we want to recover from a loss in some window of time,
we can manually run the glue job manually in AWS glue,
passing the boundaries of the window:

```
dp.from = "2023-07-27T09:34:00"
dp.to = "2023-07-28T10:04:00"
```